Date: Fri, 2 Jan 2009 02:01:16 -0800 From: "E. Wing" Subject: [SDL] Submission: ImageIO backend for SDL_Image on Mac OS X
I have a submission for SDL_image. I have implemented an ImageIO
backend for Mac OS X.

I've uploaded the code here:
http://www.assembla.com/spaces/SDL_Image_ImageIO

A Mercurial repository and a raw tar-ball can be found there.


The existing SDL_image implementation relies on 3rd-party libjpeg,
libpng, libgif, and libtiff which is a pain to build and maintain. In
the Mac release, we currently statically link all these libraries into
the framework. But Mac OS X already comes with these libraries, but
they are hidden away underneath their ImageIO framework. This
submission uses ImageIO to load these image formats.

To further minimize duplicated code, I also moved the bmp and tga
loader to ImageIO. Incidentally, Apple seems to have an implementation
that can detect if a file is a TGA.


ImageIO was introduced in Tiger 10.4. So this submission is intended
to go with the SDL_image that will accompany the SDL 1.3 release.

This code works so far for everything I've thrown at it (both PowerPC
and Intel). But I don't have a comprehensive library of test images.
So this could would benefit greatly from more scrutiny and testing.

--HG--
extra : convert_revision : svn%3Ac70aab31-4412-0410-b14c-859654838e24/trunk%404336