Jamie Knight to slouken
Hi Sam,

I'm working on a cross-platform casual game and we're currently evaluating using SDL 1.3 to deploy it for the iPhone.

Some of our textures are currently in tga files current SDL_image built from subversion fails to load on the iPhone.  I think I've tracked this down and it is basically because while the ImageIO backend supports TGA, UIImage does not and the fallback code path to use the old RW_ops loaders is broken AND the tga loader is compiled out for ImageIO and UIImage.